+++ lib/src/ats/ats.dart
@@ -21,6 +21,7 @@ part 'src/ns_license.dart';
 part 'src/transport_information.dart';
 part 'src/command_result.dart';
 part 'src/execute_load.dart';
+part 'src/booking.dart';
 
 // ATS Reception Models
 part 'src/reception/company_information.dart';
@@ -62,6 +63,8 @@ part 'src/converters/cf_fuel_type.dart';
 part 'src/converters/reception_status.dart';
 part 'src/converters/params_form_modality.dart';
 part 'src/converters/replenishment_or_withdraw.dart';
+part 'src/converters/booking_status.dart';
+part 'src/converters/booking_type.dart';
 
 // ATS Enums
 part 'src/enums/fuel_sub_type.dart';
@@ -75,6 +78,8 @@ part 'src/enums/reception_status.dart';
 part 'src/enums/monitor_card_type.dart';
 part 'src/enums/params_form_modality.dart';
 part 'src/enums/replenishment_or_withdraw.dart';
+part 'src/enums/booking_status.dart';
+part 'src/enums/booking_type.dart';
 
 // Manifests
 part 'src/manifest/kind.dart';