+++ lib/src/ats/src/converters/booking_type.dart
@@ -1,0 +1,17 @@
+part of '../../ats.dart';
+
+class AtsBookingTypeOrNullConverter implements JsonConverter<AtsBookingType?, String?> {
+  const AtsBookingTypeOrNullConverter();
+  @override
+  AtsBookingType? fromJson(String? json) {
+    if (json == null) {
+      return null;
+    }
+    return AtsBookingType.fromJson(json);
+  }
+
+  @override
+  String? toJson(AtsBookingType? object) {
+    return object?.toJson();
+  }
+}